Fix qwen_image_edit_2509_white_film_to_rendering: cast transformer to target dtype after fuse_lora

+++ qwen_image_edit_2509_white_film_to_rendering/pytorch/loader.py
@@ -80,7 +80,7 @@ class ModelLoader(ForgeModel):
         )
         pipeline.fuse_lora()
 
-        self._transformer = pipeline.transformer
+        self._transformer = pipeline.transformer.to(dtype)
         self._transformer.eval()
         return self._transformer
